Move platforms to agent platform folder
