+++ test/new-e2e/tests/agent-platform/platforms/platforms.go
